refactor: split main.ts into separate files
